+++ src/writer.rs
@@ -6,7 +6,10 @@ use crate::encoding::UTF8_BOM;
 use crate::errors::Result;
 use crate::events::{attributes::Attribute, BytesCData, BytesStart, BytesText, Event};
 
-/// XML writer. Writes XML [`Event`]s to a [`std::io::Write`] implementor.
+#[cfg(feature = "async-tokio")]
+mod async_tokio;
+
+/// XML writer. Writes XML [`Event`]s to a [`std::io::Write`] or [`tokio::io::AsyncWrite`] implementor.
 ///
 /// # Examples
 ///
@@ -53,26 +56,18 @@ use crate::events::{attributes::Attribute, BytesCData, BytesStart, BytesText, Ev
 /// assert_eq!(result, expected.as_bytes());
 /// ```
 #[derive(Clone)]
-pub struct Writer<W: Write> {
+pub struct Writer<W> {
     /// underlying writer
     writer: W,
     indent: Option<Indentation>,
 }
 
-impl<W: Write> Writer<W> {
+impl<W> Writer<W> {
     /// Creates a `Writer` from a generic writer.
     pub fn new(inner: W) -> Writer<W> {
         Writer {
             writer: inner,
             indent: None,
-        }
-    }
-
-    /// Creates a `Writer` with configured whitespace indents from a generic writer.
-    pub fn new_with_indent(inner: W, indent_char: u8, indent_size: usize) -> Writer<W> {
-        Writer {
-            writer: inner,
-            indent: Some(Indentation::new(indent_char, indent_size)),
         }
     }
 
@@ -89,6 +84,16 @@ impl<W: Write> Writer<W> {
     /// Get a reference to the underlying writer.
     pub fn get_ref(&self) -> &W {
         &self.writer
+    }
+}
+
+impl<W: Write> Writer<W> {
+    /// Creates a `Writer` with configured whitespace indents from a generic writer.
+    pub fn new_with_indent(inner: W, indent_char: u8, indent_size: usize) -> Writer<W> {
+        Writer {
+            writer: inner,
+            indent: Some(Indentation::new(indent_char, indent_size)),
+        }
     }
 
     /// Write a [Byte-Order-Mark] character to the document.